+++ SistemaHotelero/src/main/java/Hoteleria/vista/Mantenimiento_Pisos.java
@@ -5,6 +5,7 @@
  */
 package Hoteleria.vista;
 
+import Hoteleria.datos.ConexionHoteleria;
 import java.io.File;
 import java.util.List;
 import javax.swing.ButtonGroup;
@@ -12,8 +13,16 @@ import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
 import Hoteleria.datos.PisosDAO;
 import Hoteleria.dominio.Pisos;
+import java.sql.Connection;
+import java.util.HashMap;
+import java.util.Map;
 import javax.swing.JLabel;
 import javax.swing.table.DefaultTableCellRenderer;
+import net.sf.jasperreports.engine.JasperCompileManager;
+import net.sf.jasperreports.engine.JasperFillManager;
+import net.sf.jasperreports.engine.JasperPrint;
+import net.sf.jasperreports.engine.JasperReport;
+import net.sf.jasperreports.view.JasperViewer;
 
 /**
  *
@@ -46,18 +55,18 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
         modelo1.addColumn("Descripción");
         modelo1.addColumn("Estado");
         jTable.setModel(modelo1);
-        
+
         centro.setHorizontalAlignment(JLabel.CENTER);
         jTable.getColumnModel().getColumn(0).setCellRenderer(centro);
         jTable.getColumnModel().getColumn(1).setCellRenderer(centro);
         jTable.getColumnModel().getColumn(2).setCellRenderer(centro);
         jTable.getColumnModel().getColumn(3).setCellRenderer(centro);
-        
-        jTable.getColumnModel().getColumn(0).setPreferredWidth(25);        
+
+        jTable.getColumnModel().getColumn(0).setPreferredWidth(25);
         jTable.getColumnModel().getColumn(1).setPreferredWidth(100);
         jTable.getColumnModel().getColumn(2).setPreferredWidth(150);
         jTable.getColumnModel().getColumn(3).setPreferredWidth(25);
-        
+
         String datos[] = new String[4];
         PisosDAO pisosDAO = new PisosDAO();
         List<Pisos> pisos = pisosDAO.select();
@@ -116,6 +125,7 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
         btnLimpiar = new javax.swing.JButton();
         btnAyuda = new javax.swing.JButton();
         btnRadioVacio = new javax.swing.JRadioButton();
+        btnReporte = new javax.swing.JButton();
 
         setClosable(true);
         setIconifiable(true);
@@ -177,15 +187,15 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(21, 21, 21)
+                .addContainerGap()
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(jLabel2)
                     .addComponent(txtBuscar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(btnBuscar, javax.swing.GroupLayout.PREFERRED_SIZE, 20, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(jButton1))
-                .addGap(18, 18, 18)
-                .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE)
-                .addContainerGap())
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jScrollPane2, javax.swing.GroupLayout.DEFAULT_SIZE, 251, Short.MAX_VALUE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
         jPanel2.setBorder(javax.swing.BorderFactory.createTitledBorder("Agregar Número de Piso"));
@@ -249,7 +259,7 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
         });
 
         btnAyuda.setFont(new java.awt.Font("Times New Roman", 1, 12)); // NOI18N
-        btnAyuda.setText("Ayuda");
+        btnAyuda.setText("?");
         btnAyuda.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 btnAyudaActionPerformed(evt);
@@ -259,6 +269,14 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
         btnRadioVacio.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 btnRadioVacioActionPerformed(evt);
+            }
+        });
+
+        btnReporte.setFont(new java.awt.Font("Times New Roman", 1, 12)); // NOI18N
+        btnReporte.setText("Reporte");
+        btnReporte.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnReporteActionPerformed(evt);
             }
         });
 
@@ -276,7 +294,9 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
                         .addGap(18, 18, 18)
                         .addComponent(btnLimpiar)
                         .addGap(18, 18, 18)
-                        .addComponent(btnAyuda, javax.swing.GroupLayout.PREFERRED_SIZE, 82, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addComponent(btnReporte)
+                        .addGap(18, 18, 18)
+                        .addComponent(btnAyuda))
                     .addGroup(jPanel2Layout.createSequentialGroup()
                         .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                             .addComponent(jLabel3)
@@ -296,37 +316,38 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
                                 .addComponent(txtNumeroDePiso, javax.swing.GroupLayout.Alignment.LEADING)
                                 .addComponent(txtCantidadHabitaciones, javax.swing.GroupLayout.Alignment.LEADING)))
                         .addGap(3, 3, 3)))
-                .addGap(30, 30, 30))
+                .addContainerGap(30, Short.MAX_VALUE))
         );
         jPanel2Layout.setVerticalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
-                .addGap(30, 30, 30)
+                .addContainerGap(30, Short.MAX_VALUE)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(jLabel1)
                     .addComponent(txtNumeroDePiso, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
+                .addGap(18, 18, Short.MAX_VALUE)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(jLabel3)
                     .addComponent(txtCantidadHabitaciones, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 21, Short.MAX_VALUE)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(jLabel4)
                     .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 21, Short.MAX_VALUE)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                         .addComponent(jLabel6)
                         .addComponent(btnRadioInactivo)
                         .addComponent(btnRadioActivo))
                     .addComponent(btnRadioVacio))
-                .addGap(20, 20, 20)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 20, Short.MAX_VALUE)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(btnLimpiar)
                     .addComponent(btnAgregar)
                     .addComponent(btnModificar)
-                    .addComponent(btnAyuda))
-                .addGap(30, 30, 30))
+                    .addComponent(btnAyuda)
+                    .addComponent(btnReporte))
+                .addContainerGap(30, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
@@ -334,11 +355,11 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(18, 18, 18)
                 .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 10, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(lblModulo))
         );
         layout.setVerticalGroup(
@@ -348,11 +369,11 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
                 .addComponent(lblModulo)
                 .addGap(292, 292, 292))
             .addGroup(layout.createSequentialGroup()
-                .addContainerGap(23, Short.MAX_VALUE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
                     .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                     .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap(25, Short.MAX_VALUE))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
         pack();
@@ -407,13 +428,14 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
                 if (btnRadioActivo.isSelected()) {
                     pisosModificar.setEstado_De_Piso(1);
                 }
-//                if (btnRadioVacaciones.isSelected()) {
-//                    pisosModificar.setEstado_De_Piso(2);
-//                }
-                btnRadioVacio.setSelected(true);
-                limpiar();
-                pisosDAO.update(pisosModificar);
-                JOptionPane.showMessageDialog(null, "Modificación Exitosa.");
+                int confirmar = JOptionPane.showConfirmDialog(this, "¿Desea modificar el registro?", "Modificar", JOptionPane.YES_NO_OPTION);
+
+                if (confirmar == 0) {
+                    btnRadioVacio.setSelected(true);
+                    limpiar();
+                    pisosDAO.update(pisosModificar);
+                    JOptionPane.showMessageDialog(null, "Modificación Exitosa.");
+                }
             }
         } else {
             JOptionPane.showMessageDialog(null, "Todos los campos tienen que estar llenos");
@@ -493,17 +515,43 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
     }//GEN-LAST:event_btnRadioVacioActionPerformed
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
-        
+
         Pisos pisosEliminar = new Pisos();
         PisosDAO pisosDAO = new PisosDAO();
 
         pisosEliminar.setId_Numero_De_Piso(Integer.parseInt(txtBuscar.getText()));
-        pisosDAO.delete(pisosEliminar);
-        JOptionPane.showMessageDialog(null, "Registro Eliminado.");
-        btnRadioVacio.setSelected(true);
-        tabla();
-        limpiar();
+        int confirmar = JOptionPane.showConfirmDialog(this, "¿Desea eliminar el registro?", "Eliminar", JOptionPane.YES_NO_OPTION);
+
+        if (confirmar == 0) {
+            pisosDAO.delete(pisosEliminar);
+            JOptionPane.showMessageDialog(null, "Registro Eliminado.");
+            btnRadioVacio.setSelected(true);
+            tabla();
+            limpiar();
+        }
     }//GEN-LAST:event_jButton1ActionPerformed
+
+    private Connection connection = null;
+
+    private void btnReporteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnReporteActionPerformed
+
+        Map p = new HashMap();
+        JasperReport report;
+        JasperPrint print;
+
+        try {
+            connection = ConexionHoteleria.getConnection();
+            report = JasperCompileManager.compileReport(new File("").getAbsolutePath()
+                    + "/src/main/java/Hoteleria/reportes/ReportePisos.jrxml");
+            print = JasperFillManager.fillReport(report, p, connection);
+            JasperViewer view = new JasperViewer(print, false);
+            view.setTitle("Reporte de Pisos");
+            view.setVisible(true);
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }//GEN-LAST:event_btnReporteActionPerformed
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
@@ -515,6 +563,7 @@ public class Mantenimiento_Pisos extends javax.swing.JInternalFrame {
     private javax.swing.JRadioButton btnRadioActivo;
     private javax.swing.JRadioButton btnRadioInactivo;
     private javax.swing.JRadioButton btnRadioVacio;
+    private javax.swing.JButton btnReporte;
     private javax.swing.JButton jButton1;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
